requirements.txt
ultralytics==8.0.0
flwr==1.4.0
torch>=1.8.0
torchvision>=0.9.0
numpy>=1.19.5
opencv-python>=4.5.1
Pillow>=8.3.1
matplotlib>=3.4.3
seaborn>=0.11.0
pandas>=1.2.4
tqdm>=4.62.2
pyyaml>=5.4.1
